A decompilation loops support patch
I implemented a method crash_analyzer::Decompiler::updateBackwardBranch which updates a backward jump. It is assumed that the destination of a jump is a loop condition basic block (for/while loop)  or a loop body beginning (do-while loop).
It splits the target MachineBasicBlock into 2 blocks (if a target MachineInstr has a predecessor) and updates a branch instruction so that it points to a destination basic block.
The method crash_analyzer::Decompiler::splitMachineBasicBlock is used instead of the method MachineBasicBlock::splitAt because the method call MachineBasicBlock::normalizeSuccProbs() isn't necessary because it converts unknown probabilities into known ones.

I also added a loop decompilation test.
I added analysis tests. The order of traversing basic blocks is same when testing decompiled .mir code and .mir generated by clang.