A decompilation improvement
I improved a creation of a new machine basic block when the current instruction
is a destination of some branch that was decompiled. If a previously decompiled
instruction was a branch, there is no need to create a new machine basic block
because it was created (by the code that is located later, before the end of the
function).

I also corrected a test related to this decompilation stuff.